Move production order test into package
